requirements.txt
asgiref==3.5.2
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.6.0
Django==4.0.6
django-ninja==0.19.0
fonttools==4.34.4
html5lib==1.1
Pillow==9.2.0
pycparser==2.21
pydantic==1.9.1
pydyf==0.2.0
pyphen==0.12.0
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
typing_extensions==4.3.0
tzdata==2022.1
weasyprint==56.0
webencodings==0.5.1
zopfli==0.2.1
